+++ packages/pn-personafisica-login/src/services/configuration.service.ts
@@ -17,7 +17,7 @@ export interface LoginConfiguration {
   IS_SMART_APP_BANNER_ENABLED?: boolean;
   ACCESSIBILITY_LINK: string;
   ONE_IDENTITY_LOGIN_ENABLED: boolean;
-  ONE_IDENTITY_CLIENT_ID: string;
+  ONE_IDENTITY_CLIENT_ID?: string;
   ONE_IDENTITY_BASE_URL: string;
 }
 
@@ -42,7 +42,14 @@ class LoginConfigurationValidator extends Validator<LoginConfiguration> {
     this.ruleFor('IS_SMART_APP_BANNER_ENABLED').isBoolean();
     this.ruleFor('ACCESSIBILITY_LINK').isString().isRequired();
     this.ruleFor('ONE_IDENTITY_LOGIN_ENABLED').isBoolean();
-    this.ruleFor('ONE_IDENTITY_CLIENT_ID').isString().isRequired();
+    this.ruleFor('ONE_IDENTITY_CLIENT_ID')
+      .isString()
+      .customValidator((value, model) => {
+        if (model.ONE_IDENTITY_LOGIN_ENABLED && !value) {
+          return 'ONE_IDENTITY_CLIENT_ID is required when ONE_IDENTITY_LOGIN_ENABLED is true';
+        }
+        return null;
+      });
     this.ruleFor('ONE_IDENTITY_BASE_URL').isString().isRequired();
   }
 }