+++ packages/pn-personagiuridica-webapp/src/setupTests.tsx
@@ -42,6 +42,7 @@ beforeAll(async () => {
     DOWNTIME_EXAMPLE_LINK: 'https://www.example.com',
     SELFCARE_CDN_URL: 'https://cdn.selfcare.pagopa.it',
     ACCESSIBILITY_LINK: 'https://accessibility-link.it',
+    NOTIFICATION_CANCELLED_HELP_LINK: 'https://fake.notification-cancelled-help.pagopa.it',
   });
   initStore(false);
   initAxiosClients();